requirements.txt
# --- Web/API ---
fastapi>=0.115
uvicorn[standard]>=0.30

# --- Async / HTTP / DB ---
asyncpg>=0.30
aiohttp>=3.9
requests>=2.32

# --- Numeric / Data ---
numpy>=2.0
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4
joblib>=1.3
psutil>=5.9
matplotlib>=3.8
seaborn>=0.13
python-dateutil>=2.8.2
pytz>=2024.1

# --- Industrial / IoT ---
pymodbus>=3.5
opcua>=0.98.13
paho-mqtt>=2.0

# --- NOTE ---
# Removed: asyncio, uuid, hashlib2 (all invalid/unneeded)
# TensorFlow is optional; see extras-ml.txt
alembic>=1.13
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

redis>=5.0


pytest>=7.4
pytest-asyncio>=0.23
httpx>=0.27
asgi-lifespan>=2.1
